requirements.txt
# Python dependencies
opencv-python>=4.8.0
kafka-python>=2.0.2
pymongo>=4.5.0
ultralytics>=8.0.0
openai-whisper>=20230314
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
